// src/services/models/InfantModel.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type InfantModel = {
    firstName?: string | null;
    lastName?: string | null;
    dateOfBirth?: string | null;
    gender?: string | null;
    lengthOfBirth?: number;
    modeOfDelivery?: string | null;
    gestationWeek?: number;
    guardianFirstName?: string | null;
    guardianLastName?: string | null;
    email?: string | null;
    phone?: string | null;
    guardianDateOfBirth?: string;
    role?: string | null;
    address?: string | null;
    alternatePhoneNumber?: string | null;
    city?: string | null;
    guardianType?: string | null;
    nextOfKinName?: string | null;
    nextOfKinAddress?: string | null;
    nextOfKinPhoneNumber?: string | null;
    nextOfKinEmail?: string | null;
};